Seed admin user and sample posts on startup
