+++ 11xtesting/README.md
@@ -14,6 +14,9 @@ pnpm dev
 bun dev
 ```
 
+The app seeds a default user and two posts whenever `npm run dev` is executed.
+Use **admin**/**password123** to log in.
+
 Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
 
 You can start editing the page by modifying `app/page.tsx`. The page auto-updates as you edit the file.
